backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pymongo==4.6.0
motor==3.3.2
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1
pillow==10.1.0
opencv-python==4.8.1.78
transformers==4.35.2
torch==2.1.1
spacy==3.7.2
langdetect==1.0.9
python-dotenv==1.0.0
httpx==0.25.2
pandas==2.1.3
numpy==1.23.5
scikit-learn==1.3.2 